Atualizações Exception

+++ src/main/java/com/example/Desafio3/services/ClientService.java
@@ -4,12 +4,15 @@ package com.example.Desafio3.services;
 import com.example.Desafio3.dto.ClientDTO;
 import com.example.Desafio3.entities.Client;
 import com.example.Desafio3.repositories.ClientRepository;
+import com.example.Desafio3.services.exception.DataBaseException;
 import com.example.Desafio3.services.exception.ResourceNotFoundException;
 import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 
@@ -23,7 +26,9 @@ public class ClientService {
     private ClientRepository clientRepository;
 
     @Transactional(readOnly = true)
+
     public ClientDTO findById(Long id) {
+
         Client client = clientRepository.findById(id)
             .orElseThrow(() -> new ResourceNotFoundException("Recurso não encontrado para o ID " + id));
 
@@ -52,22 +57,32 @@ public class ClientService {
 
     @Transactional
     public ClientDTO update (Long id,ClientDTO dto){
+        try {
+            Client entity = clientRepository.getReferenceById(id);
+            entity.setName(dto.getName());
+            entity.setCpf(dto.getCpf());
+            entity.setIncome(dto.getIncome());
+            entity.setBirthDate(dto.getBirthDate());
+            entity = clientRepository.save(entity);
 
-        Client entity = clientRepository.getReferenceById(id);
-        entity.setName(dto.getName());
-        entity.setCpf(dto.getCpf());
-        entity.setIncome(dto.getIncome());
-        entity.setBirthDate(dto.getBirthDate());
-
-
-        entity=clientRepository.save(entity);
-
-        return new ClientDTO(entity);
+            return new ClientDTO(entity);
+        }
+        catch(EntityNotFoundException e ){
+            throw new  ResourceNotFoundException("Recurso não encontrado");
+        }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete (Long id) {
-        clientRepository.deleteById(id);
+        if(!clientRepository.existsById(id)){
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            clientRepository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e){
+          throw new DataBaseException("Falha de integridade referencial");
+        }
     }
